+++ src/main/java/com/esotericsoftware/SpinePreview/utils/TwoColorPolygonBatch.java
@@ -707,7 +707,7 @@ public class TwoColorPolygonBatch implements PolygonBatch {
         final int vertexIndex = this.vertexIndex;
         int triangleIndex = this.triangleIndex;
         short vertex = (short) (vertexIndex / VERTEX_SIZE);
-        for (int n = triangleIndex + triangleCount; triangleIndex < n; triangleIndex += 6, vertex += 4) {
+        for (int n = triangleIndex + triangleCount; triangleIndex < n; triangleIndex += 6, vertex += (short) 4) {
             triangles[triangleIndex] = vertex;
             triangles[triangleIndex + 1] = (short) (vertex + 1);
             triangles[triangleIndex + 2] = (short) (vertex + 2);
@@ -733,7 +733,7 @@ public class TwoColorPolygonBatch implements PolygonBatch {
         final int vertexIndex = this.vertexIndex;
         int triangleIndex = this.triangleIndex;
         short vertex = (short) (vertexIndex / VERTEX_SIZE);
-        for (int n = triangleIndex + triangleCount; triangleIndex < n; triangleIndex += 6, vertex += 4) {
+        for (int n = triangleIndex + triangleCount; triangleIndex < n; triangleIndex += 6, vertex += (short) 4) {
             triangles[triangleIndex] = vertex;
             triangles[triangleIndex + 1] = (short) (vertex + 1);
             triangles[triangleIndex + 2] = (short) (vertex + 2);
